Move BucketManager from fs package to gcsx package
