+++ internal/gcsx/garbage_collect.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fs
+package gcsx
 
 import (
 	"fmt"
